Local Cosigner and Config changes

+++ cmd/horcrux/cmd/config.go
@@ -341,7 +341,7 @@ func addPeersCmd() *cobra.Command {
 				return errors.New("no new peer nodes in args")
 			}
 			diff = append(config.Config.CosignerConfig.Peers, diff...)
-			config.Config.CosignerConfig.Shares = len(diff) + 1
+			config.Config.CosignerConfig.Shares = uint8(len(diff)) + 1
 			if err := validateCosignerPeers(diff, config.Config.CosignerConfig.Shares); err != nil {
 				return err
 			}
@@ -386,7 +386,7 @@ func removePeersCmd() *cobra.Command {
 				return errors.New("cannot remove all peer nodes from config, please leave at least one")
 			}
 
-			config.Config.CosignerConfig.Shares = len(diff) + 1
+			config.Config.CosignerConfig.Shares = uint8(len(diff)) + 1
 			// If none of the peer nodes in the args are listed in the config, just continue
 			// without throwing an error, as the peer nodes in the config remain untouched.
 			if err := validateCosignerPeers(diff, config.Config.CosignerConfig.Shares); err != nil {
@@ -627,7 +627,7 @@ func validateCosignerPeers(peers []CosignerPeer, shares uint8) error {
 	// Check that exactly {num-shares}-1 peers are in the peer list, assuming
 	// the remaining peer ID is the ID the local node is configured with.
 
-	if len(peers) != shares-1 {
+	if uint8(len(peers)) != shares-1 {
 		return fmt.Errorf("incorrect number of peers. expected (%d shares - local node = %d peers)",
 			shares, shares-1)
 	}
